respuesta correcta c

+++ taller1/C.c
@@ -13,7 +13,7 @@ int main(){
     uint8_t numExcercise = 0;
 
 
-    uint8_t total = 0;
+    uint16_t total = 0;
     for(uint8_t i = 1; i <= n; i++){
         total += 5 * i;
         if(total <= maxMinutes){
